Extract shared block primitives into pages/shared_blocks.py
- Create new pages/shared_blocks.py module with color/link/id primitives
  (OpenStaxColorBlock, hex_color_block, LinkBlock, LinkInfoBlock, CTALinkBlock,
  id_config_block, gradient_config_options, gradient_block_counts)
- Remove duplicate definitions from pages/custom_blocks.py
- Update imports in pages/table_block.py, pages/models/constants.py, and
  pages/models/core.py to use pages.shared_blocks for moved items
- Update import in pages/tests/test_link_resolution.py to use new location
- Update all migrations to reference pages.shared_blocks instead of
  pages.custom_blocks for the moved block types
- Add regression test pages/tests/test_shared_blocks.py to verify the new
  module structure

This refactor breaks a circular import cycle that prevented Task 2 from adding
a content field to FAQBlock (which would embed TableBlock, which already imports
from custom_blocks).
