+++ pages/models/core.py
@@ -17,12 +17,10 @@ from salesforce.models import School
 
 from pages.custom_blocks import APIImageChooserBlock, \
     APIRichTextBlock, \
-    LinkInfoBlock, \
-    TEXT_ALIGNMENT_CHOICES, \
-    hex_color_block, \
-    gradient_config_options, \
-    gradient_block_counts, \
-    id_config_block
+    TEXT_ALIGNMENT_CHOICES
+
+from pages.shared_blocks import LinkInfoBlock, hex_color_block, \
+    gradient_config_options, gradient_block_counts, id_config_block
 
 
 from .constants import BODY_BLOCKS
